+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * allow selecting slaves with constraints (#6)
 * allow to specify mesos role (#5)
 * find mesos master with zookeeper leader detector (#9)
+* download files from slave (#13)
 
 ## v0.1.0